[cueweb] Move poller, polish subscription utilities, fix redundant re-render
