+++ cueweb/app/utils/use_job_subscriptions.ts
@@ -15,7 +15,6 @@ export function useJobSubscriptions() {
   const [store, setStore] = useState<SubscriptionStore>(getSubscriptions);
 
   useEffect(() => {
-    setStore(getSubscriptions());
     return subscribeToChanges(() => setStore(getSubscriptions()));
   }, []);
